[fix] xp 타입을 Long으로 지정

+++ src/main/kotlin/com/meokq/api/user/response/CustomerResp.kt
@@ -2,7 +2,6 @@ package com.meokq.api.user.response
 
 import com.meokq.api.user.enums.UserStatus
 import com.meokq.api.user.model.Customer
-import com.meokq.api.user.request.CustomerXpReq
 import io.swagger.v3.oas.annotations.media.Schema
 
 @Schema(name = "Customer-Response")
@@ -11,7 +10,7 @@ data class CustomerResp(
     val nickname: String?,
     val couponCount: Long,
     val completeChallengeCount: Long,
-    val xpPoint: Int
+    val xpPoint: Long
 ){
     constructor(model : Customer, couponCount: Long, challengeCount: Long) : this(
         status = model.status,
